% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortPlots.R
\name{getBalanceScatter}
\alias{getBalanceScatter}
\title{Plot Covariate Balance}
\usage{
getBalanceScatter(
  balance,
  filePath = TRUE,
  target_id,
  comparator_id,
  outcome_id,
  analysis_id
)
}
\arguments{
\item{balance}{either a valid csv path, dataframe or tibble. If using a csv path
filePath must be true}

\item{filePath}{a logic toggle indicating whether a file path is used for balance}

\item{target_id}{the target id to extract from the aggregate file}

\item{comparator_id}{the comparator id to extract from the aggregate file}

\item{outcome_id}{the outcome id to extract from the aggregate file}

\item{analysis_id}{the analysis id to exctract from the aggregate file}
}
\value{
a ggplot object of the covariate balance scatter plot
}
\description{
This function takes the covariate balance and shows the balance scatter plot
}
